+++ packages/devac-cli/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @devac/cli
+# @pietgk/devac-cli
 
 ## 0.1.1
 
 ### Patch Changes
 
 - Updated dependencies [144f370]
-  - @devac/core@0.1.1
+  - @pietgk/devac-core@0.1.1